requirements.txt
binance~=0.3
python-dotenv==0.21.0